fix: more utils functions
